Update All Libraries to More Recent Version
Following Libraries (Incl. JS, CSS and other resources) have been
updated and tested as they were toooo old:
- Almond 0.2.0 => 0.2.9
- Backbone 0.9.10 => 1.2.1
- Backbone.validateAll 0.1.0 => 0.2.0
- Backbone.Marionette 1.0.0-rc6 => 2.4.2
- Bootstrap 2.2.1 => 3.3.5
- Handlebars => 3.0.3
- Handlebars 0.4.0 => 2.0.0
- jQuery 1.8.3=> 1.11.1
- jQuery Mobile => 1.4.0
- jQuery UI 1.9.2 => 1.10.4
- Lodash 0.10.0 => 2.4.1
- RequireJS & r.js 2.1.2=> 2.1.11
- RequireJS text 2.0.0 => 2.0.7

Please notice that with the update of handlerbars and hbs, the i18n
feature was completely removed. As SlexAxton is working on implementing
Build Time Compilation for i18n in
https://github.com/SlexAxton/messageformat.js, I temporarily demised
relevant examples in the boilertemplate.
Moreover with the majot update of Bootstrap some styles within the
index.html may change.